Add HighScorerTrial as custom scoring algorithm implementation
Sadly, this feels kind of crude (I'd much rather group the scoring by
item type) and doesn't encompass some item archetypes yet
